requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-Sock==0.7.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.1
requests==2.31.0
eventlet==0.35.2
Werkzeug==3.0.1
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3
urllib3==2.0.7
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
SQLAlchemy==2.0.27
alembic==1.12.1
psutil==5.9.6 
